collector/requirements.txt
websockets
psycopg2-binary
pandas
requests
